requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
azure-functions==1.12.0
azure-identity==1.12.0
cffi==1.17.1
charset-normalizer==2.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
lxml==4.9.1
msal==1.19.0
msal-extensions==1.1.0
multidict==6.1.0
numpy==1.24.1
Office365-REST-Python-Client==2.3.11
openpyxl==3.0.10
packaging==24.1
pandas==1.5.3
portalocker==2.10.1
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.2
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
XlsxWriter==3.0.2
yarl==1.17.1
